etl-job/pyproject.toml
[project]
name = "etl-job"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.11.18",
    "azure-identity>=1.21.0",
    "azure-keyvault-secrets>=4.9.0",
    "azure-servicebus>=7.14.2",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.9.1",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "ruff>=0.11.8",
]
